requirements.txt
line-bot-sdk
flask
tensorflow
keras
gensim
numpy
jieba